generateImage deprecated
